Add search functionality to link list
This commit adds a search field to the link table. The search functionality was added in LinkTable.php by modifying the render() method to perform a wildcard search on either the 'url' or 'slug' fields whenever the 'search' property is not empty. For clean usage, it makes sure to import the Link model at the top of the file. The structure of the Link table in link-table.blade.php and links/index.blade.php was also slightly modified to better accommodate the new search field.