Update 'Links' references to 'Search' and add Spotlight package
In this commit, all references to 'Links' in multiple files were changed to 'Search' which provides a more accurate description of the functionality. A new Spotlight package was added in composer.json file to enhance search capabilities in the Laravel application. The Spotlight package adds a Livewire component that provides Spotlight/Alfred-like functionality to the application. This change is to improve user search experience.